+++ C++/IRPmodel/IRPmodel/CustomerVRPDB.h
@@ -6,7 +6,7 @@
 class CustomerVRPDB	: public CustomerDB
 {
 public:
-	CustomerVRPDB(vector <vector<int>> &Demand, vector <Customer *> & );
+	CustomerVRPDB(vector <vector<double>> &Demand, vector <Customer *> & );
 	int getnCustomers();
 	int getDemand(int id, int indicator);
 	CustomerVRP * getCustomer(int);
